# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SysMap Social Network - OpenAPI 3.0",
    "description": "Backend of the Social Network project of SysMap fullstack bootcamp.",
    "contact": {
      "email": "kellyplcastelo@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000/v1"
    }
  ],
...
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": ["users"],
        "summary": "Finds logged user",
        "description": "Finds logged user",
        "operationId": "findLoggedUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "put": {
        "tags": ["users"],
        "summary": "Updates logged user's username",
        "description": "Updates username",
        "operationId": "updateUser",
        "requestBody": {
          "description": "Updates logged in username",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "203": {
            "description": "UNDOCUMENTED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "delete": {
        "tags": ["users"],
        "summary": "Deletes logged user",
        "description": "deletes logged user",
        "operationId": "deleteUser",
        "responses": {
          "204": {
            "description": "NO CONTENT"
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
...
    "/security/register": {
      "post": {
        "tags": ["security"],
        "summary": "Creates a new user",
        "description": "Creates a new user",
        "operationId": "registerUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CREATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorArrayResponse"
                }
              }
            }
          }
        }
      }
    },
    "/security/login": {
      "post": {
        "tags": ["security"],
        "summary": "Logs in an user",
        "description": "Logs in an user",
        "operationId": "logInUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "minLength": 2
          },
          "password": {
            "type": "string",
            "minLength": 2
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiJ9.YmFyYmE.2SRqT4_DuKueHrzf6pFTu7EUZx1QSLoJhoT_kWT-0z4"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "example": "Kelly"
          },
          "user": {
            "type": "string",
            "minLength": 2,
            "example": "Castelo"
          },
          "password": {
            "type": "string",
            "minLength": 2,
            "example": "randompassoword123"
          }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "637c9bb1fceec2bc7d8d9249"
          },
          "user": {
            "type": "string",
            "example": "Castelo"
          }
        }
      },
      "ErrorArrayResponse": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "example": "/v1/path/action"
          },
          "status": {
            "type": "integer",
            "example": "400"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "Path `user` is required"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "example": "/v1/path/action"
          },
          "status": {
            "type": "integer"
          },
          "errors": {
            "type": "string"
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "637bb9d4da4e28ea8606b7bf"
          },
          "user": {
            "type": "string",
            "example": "Castelo"
          },
          "profile": {
            "type": "string",
            "example": "637bb9d4da4e28ea8606b7c1"
          }
        }
      },
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "Castelo"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}